% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{cleansing_umlaute}
\alias{cleansing_umlaute}
\title{Umlaute Entfernen}
\usage{
cleansing_umlaute(x, ...)
}
\arguments{
\item{x}{string}
}
\description{
Funktion `cleansing_umlaute()` entfernt stoerende Umlaute, unten stehende Liste ggf. erweitern.
}
