% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTML.R
\name{HTML_TABLE}
\alias{HTML_TABLE}
\title{html-table}
\usage{
HTML_TABLE(df, ...)
}
\arguments{
\item{df}{data.frame}
}
\description{
Experimental
}
\examples{

 HTML_TABLE( data.frame(
    Item = c("A", "B", "Ö"),
    x_m = c(45.47, 256.14, 14.47),
    x_sd = c(1.43, 25.264, 5.423),
    y_m = c(35.42, 261.94, 24.43),
    y_sd = c(3.44, 25.25, 5.42)

)

}
