% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output-GT.R, R/Output-other.R, R/Output.R,
%   R/Output_kable.R
\name{Output.gt_tbl}
\alias{Output.gt_tbl}
\alias{Output.html}
\alias{gt_split_header}
\alias{Output.gtsummary}
\alias{Output.sjTable}
\alias{Output}
\alias{Output2}
\alias{Output.character}
\alias{Output.data.frame}
\alias{Output.list}
\alias{Output.default}
\alias{Output_kable}
\alias{Output_kable.list}
\alias{Output_kable.default}
\title{Ausgabe von HTML, Knit oder Text}
\usage{
\method{Output}{gt_tbl}(
  x,
  caption = NULL,
  note = NULL,
  split_header = TRUE,
  na.rm = TRUE,
  output = which_output(),
  ...
)

\method{Output}{html}(x, ..., output = which_output())

gt_split_header(x, na.rm = TRUE, missing_text = "", ...)

\method{Output}{gtsummary}(x, caption = "", note = "", ...)

\method{Output}{sjTable}(x, caption = NULL, note = NULL, output = which_output(), ...)

Output(x, ...)

Output2(x, ..., note = NULL, header = NULL, output = which_output())

\method{Output}{character}(x, ...)

\method{Output}{data.frame}(
  x,
  caption = NULL,
  note = NULL,
  col.names = NULL,
  header = col.names,
  select = NA,
  output = which_output(),
  split_header = TRUE,
  css.table = "padding-left: .5em; padding-right: .2em;",
  css.cell = "padding-left: .5em; padding-right: .2em;",
  booktabs = TRUE,
  latex_options = c("hold_position"),
  linesep = "",
  align = "l",
  rgroup = attr(x, "rgroup", TRUE),
  n.rgroup = attr(x, "n.rgroup", TRUE),
  ...
)

\method{Output}{list}(x, caption = NULL, note = NULL, output = which_output(), ...)

\method{Output}{default}(x, ...)

Output_kable(x, ...)

\method{Output_kable}{list}(x, ...)

\method{Output_kable}{default}(
  x,
  caption = NULL,
  col.names = colnames(x),
  booktabs = TRUE,
  latex_options = c("hold_position"),
  note = NULL,
  output = which_output(),
  ...
)
}
\arguments{
\item{x}{dataframe}

\item{caption, note}{Ueberschrift Fussnote}

\item{split_header}{in GT Spaltenüberschrift spliten}

\item{na.rm, missing_text}{replace NA}

\item{output}{welcher output, text, html, markdown}

\item{...}{nicht benutzt}

\item{header, col.names}{altlast beides das selbe Kopf Zeile}

\item{select}{auswahl von Spalten}

\item{css.table, css.cell, align}{htmlTable
padding-left: .5em; padding-right: .2em;}

\item{booktabs, latex_options}{kable an Latex     latex_options = c("hold_position"),
latex_options = "scale_down"}

\item{linesep}{linesep = ""  linesep = c("", "", "\\midrule")}

\item{rgroup, n.rgroup}{an htmlTable rgroup = c("A", "B", "C")
n.rgroup wie oft kommt A b und C vor.
Die Position kann über   rgroup = c("A"=1, "B"=3, "C" =6 ) eingestellt werden}
}
\value{
gt_tbl, list

HTML oder Textausgabe

input
}
\description{
Output mit gt() interne übergabe an gt_split_header()

gt_split_header: GT Tabel header spliten

HTML :  htmlTable::htmlTable

Output.iste: einzeln in einen data.frame transformieren

Output mit  knitr::kable Funftioniert nur wenn options(knitr.table.format = "latex")
gesetzt wird (das gilt nur fuer Pdf)
}
\details{
markdown: knitr::kable -> kableExtra::kable_styling

Text: knitr::kable

Word: knitr::kable(x, format = "pandoc")

Output_word() flextable::regulartable
}
\examples{

require(gt)
# require(stp25output2)

tab <- data.frame(
  Item = c("Aaa", "Bbb", "Ööö"),
  x_m = c(45.47, 26.14, 14.47),
  x_sd = c(1.43, 2.26, 5.42),
  y_m = c(35.42, 61.94, 24.43),
  y_sd = c(3.44, 5.25, 5.42)
)

tab |>
  gt() |> Output()

tab |>
  gt() |>
  as_raw_html() |>
  Output()
\dontrun{
library(gtsummary)
 tbl_summary_ex1 <-
trial |>
  select(age, grade, response) |>
  tbl_summary() |>
  Output()
}
 \dontrun{

library(sjPlot)
library(sjmisc)
library(sjlabelled)

data("efc")
efc <- as_factor(efc, c161sex, c172code)

m1 <- lm(barthtot ~ c160age + c12hour + c161sex + c172code, data = efc)
m2 <- lm(neg_c_7 ~ c160age + c12hour + c161sex + e17age, data = efc)

sjPlot::tab_model(m1, m2) |>
  Output("Side-by-side Regression Models")
}


df1 <- data.frame(
term = c("A", "B", "C", "D"),
n = c(23, 14, 56, 2),
m = c("4.7 (2.4)", "4.1 (2.3)", "8.9 (3.6)", NA)
)


df2 <- data.frame(
  term = c("A", "B", "C", "D"),
  G1_k_n = c(23, 14, 56, 2),
  G1_k_m = c("4.7 (2.4)", "4.1 (2.3)", "8.9 (3.6)", NA),
  G2_n = c(33, 35, 78, 21),
  G2_m = c("4.9 (2.7)", "4.7 (2.5)", "4.1 (5.6)", "4.2 (5.6)")
  )
#+ df-default ,  results='asis'
df1 |> Output()

#+ df-false ,  results='asis'
df1 |> Output()


#+ df-html
df1 |> Output(output="html")

#+ df-text ,  results='asis'
df1 |> Output(output="text")

#+ df-word ,  results='asis'
 # df1 |> Output(output="word")


#+ df-mark
df1 |> Output(output="markdown")

# df1 |>  Output(linesep = c("", "", "\\midrule"))

}
