% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_word.R
\name{Output_word}
\alias{Output_word}
\title{Output nach docx}
\usage{
Output_word(
  x,
  caption = NULL,
  note = NULL,
  output = "docx",
  split_header = TRUE,
  wrap = get_opt("table", "wrap"),
  wrap_result = get_opt("table", "wrap_result"),
  ...
)
}
\arguments{
\item{x}{data.frame}

\item{caption, note, output, split_header, wrap, wrap_result}{Ueberschrift usw}

\item{...}{not used}
}
\value{
flextable
}
\description{
Geht nur direkt  und arbeitet mit flextable
}
\details{
asis_output() function only works in top-level R expressions

Wine Alternative ist die  gt-Bibilothek
}
