% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagebreak.R
\name{pagebreak}
\alias{pagebreak}
\title{pagebreak}
\usage{
pagebreak()
}
\value{
nichts
}
\description{
Neue Seite bei copy and paste in Word.
Geht nicht mehr ??
}
\details{
Copy/Paste geht nicht mehr  aber das öffnen der HTML-Seite mit Word funktioniert.
}
\examples{
 \dontrun{
# in MD-Files
# `r pagebreak()`
}
}
