% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projekt.R
\name{Projekt}
\alias{Projekt}
\alias{Rmd_Start}
\alias{End}
\alias{Stop}
\title{Projekt}
\usage{
Projekt(myformat = "", Projektname = "Demo", datum = date())

Rmd_Start(myformat, Projektname, datum)

End(
  browser = "iexplore",
  output = options()$prompt[1] == "HTML> ",
  file = HTMLGetFile(),
  ...
)

Stop()
}
\arguments{
\item{myformat}{HTML, Spin, Knit, Rpres oder Text. Spin ist knitr wobei die Ausgabe der Tabellen mit html erfolgt}

\item{Projektname}{Bezeichnung des Projektes (gilt auch fuer die HTML Seite)}

\item{datum}{Datum zur Dokumentation}

\item{browser}{Ie oder Chrome}

\item{output, file}{intern}
}
\value{
html-Pfad oder Projektname als Text
}
\description{
Ein R-File ruft die Interne Funktion Projekt_html() auf und ein Rmd-File Rmd_Start().

Einstellungen fuer .Rmd files hier werden keine Folder erstellt

\subsection{End}{
Zuruecksetzen der Einstellungen und Aufruf des Browsers browser = getOption("browser")}
}
\details{
Start eines neuen Prpjekts
n.
}
\examples{

\dontrun{

 Projekt()


 # Auswertung


 End()

}

}
