% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{which_output}
\alias{get_opt}
\alias{set_opt}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{stp25settings}{\code{\link[stp25settings:set_opt]{get_opt}}, \code{\link[stp25settings]{set_opt}}, \code{\link[stp25settings]{which_output}}}
}}

