% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_data.R
\name{report_data}
\alias{report_data}
\title{Reporting the participant data}
\usage{
report_data(data, caption = "Excluded Participants")
}
\arguments{
\item{data}{A data frame.}

\item{caption}{Tabellen Überschrift}
}
\description{
Reporting the participant data
}
\examples{

data <- data.frame(
  "Age" = c(22, 23, 54, 21, 8, 42, 18, 32, 24, 27, 45),
  "Sex" = c("Intersex", "F", "F", "M", "M", "M", "F", "F", "F", "F", "F"),
  "Gender" = c("N", "W", "W", "M", "M", "M", "W", "W", "W", "W", "W"),
  "Race" = c(
    "Black", NA, "White", "Asian", "Black", "Arab", "Black",
    "White", "Asian", "Southeast Asian", "Mixed"
  )
)

data<- rbind(data,data,data) |> stp25tools::filter2(Gender!="N")

report_data(data )

}
