% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SavePlot.R
\name{SavePlot}
\alias{SavePlot}
\title{SavePlot}
\usage{
SavePlot(
  caption = "",
  w = dev.size("in")[1],
  h = dev.size("in")[2],
  filename = "",
  save_plot = "pdf",
  output = which_output(),
  res = 72,
  out.type = "pdf"
)
}
\arguments{
\item{caption}{TextAbbildung}

\item{w, h, res}{Width, Height, an HTMLplot
default = 520, Breite der  Grafik  also zb w=8=dev.size("in")[2],
h=dev.size("in")[1] Hoehe der Grafik also zb h=8
res=72}

\item{filename}{Text}

\item{save_plot}{speichern als file = pdf esp,}

\item{output}{nur wichtig bei html}

\item{out.type}{name of the output device: can
be "pdf", or "cairo"}
}
\value{
file name
}
\description{
Speichern von Grafiken
Grafik werden  als PDF oder Windows-Metafile
speichern. Die Funktion Abb() beinhaltet den Pfad.
}
\examples{

plot(1)
SavePlot("Einfache Grafik", w=4, h=4)

#'  require(cowplot)
title <-
  cowplot::ggdraw() +
  cowplot::draw_label("FFQ DEGS 534", fontface = 'bold')

cowplot::plot_grid(
  title, NULL,
  function(x) plot(1),
  function(x) plot(1:10),
  nrow = 2,
  #labels = c("", "", "C"),
  rel_heights = c(0.05, 1 )
)
cowplot::save_plot( )
ggplot2::ggsave("test-23.pdf")
ggplot2::ggsave("test-2.eps")

resulution
w und h in pixels
save as png
}
