% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_abb.R
\name{Tab_Abb}
\alias{Tab_Abb}
\alias{Tab}
\alias{Tab_Index}
\alias{Caption}
\alias{Note}
\alias{Abb_Index}
\alias{Abb}
\title{Tabellen und Abbildungs Nummern}
\usage{
Tab(caption = "", caption.nr.prefix = get_opt("table", "caption.nr.prefix"))

Tab_Index(x = NULL)

Caption(
  caption = NULL,
  atr = NULL,
  N = NULL,
  include.n = get_opt("table", "include.n"),
  include.tabel.number = get_opt("table", "include.tabel.number")
)

Note(x = NULL, atr = NULL)

Abb_Index(x = NULL)

Abb(filename = "", caption = "", folder = get_opt("fig_folder"))
}
\arguments{
\item{caption}{Name der Tabelle/Abbildung}

\item{caption.nr.prefix}{was soll vorne stehen}

\item{x}{Text}

\item{atr}{in Caption: alternativer Text}

\item{N}{Stichprobe}

\item{include.n, include.tabel.number}{wird über get_opt gesteuert
entweder direkt include.n oder indirekt mit caotion =TRUE/FALSE}

\item{filename}{Name des Foles}

\item{folder}{Speicherort}
}
\value{
Name plus Nummer
}
\description{
Nummerierte Abbildung default Folder ist fig

Ueberschrift aus stp-Objekt:
}
\examples{
# example code

# \dontrun{
# Caption(caption, attr(x, "caption"))
# oder Note(note, attr(x, "note")) mit
# }
}
