% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output-other.R, R/Output.R, R/Output_kable.R,
%   R/Text.R
\name{Output.gtsummary}
\alias{Output.gtsummary}
\alias{Output}
\alias{Output2}
\alias{Output.character}
\alias{Output.data.frame}
\alias{Output.list}
\alias{Output.default}
\alias{Output_kable}
\alias{Output_kable.list}
\alias{Output_kable.default}
\alias{Text.default}
\alias{Text.character}
\alias{Text.data.frame}
\title{Ausgabe von HTML, Knit oder Text}
\usage{
\method{Output}{gtsummary}(x, caption = "", note = "", ...)

Output(x, ...)

Output2(x, ..., note = NULL, header = NULL, output = which_output())

\method{Output}{character}(x, ...)

\method{Output}{data.frame}(
  x,
  caption = NULL,
  note = NULL,
  header = NULL,
  select = NA,
  col.names = NULL,
  output = which_output(),
  split_header = TRUE,
  css.table = "padding-left: .5em; padding-right: .2em;",
  css.cell = "padding-left: .5em; padding-right: .2em;",
  booktabs = TRUE,
  latex_options = c("hold_position"),
  linesep = "",
  align = "l",
  rgroup = attr(x, "rgroup", TRUE),
  n.rgroup = attr(x, "n.rgroup", TRUE),
  ...
)

\method{Output}{list}(x, caption = NULL, note = NULL, output = which_output(), ...)

\method{Output}{default}(x, ...)

Output_kable(x, ...)

\method{Output_kable}{list}(x, ...)

\method{Output_kable}{default}(
  x,
  caption = NULL,
  col.names = colnames(x),
  booktabs = TRUE,
  latex_options = c("hold_position"),
  note = NULL,
  output = which_output(),
  ...
)

\method{Text}{default}(x, ...)

\method{Text}{character}(..., style = 0, char = "-", output = which_output())

\method{Text}{data.frame}(
  x,
  ...,
  style = 0,
  char = "-",
  include.levels = FALSE,
  output = which_output()
)
}
\arguments{
\item{x}{dataframe}

\item{caption, note}{Ueberschrift Fussnote}

\item{...}{nicht benutzt}

\item{header}{erste Zeile nur bei HTML}

\item{output}{welcher output, text, html, markdown}

\item{select}{auswahl von Spalten}

\item{col.names}{Output_kable:  fuer tintPdf format = "latex"}

\item{css.table, css.cell, align}{htmlTable
padding-left: .5em; padding-right: .2em;}

\item{booktabs, latex_options}{kable an Latex     latex_options = c("hold_position"),
latex_options = "scale_down"}

\item{linesep}{linesep = ""  linesep = c("", "", "\\midrule")}

\item{rgroup, n.rgroup}{an htmlTable rgroup = c("A", "B", "C")
n.rgroup wie oft kommt A b und C vor.
Die Position kann über   rgroup = c("A"=1, "B"=3, "C" =6 ) eingestellt werden}
}
\value{
HTML oder Textausgabe

nix
}
\description{
HTML :  htmlTable::htmlTable

Output.iste: einzeln in einen data.frame transformieren

Output mit  knitr::kable Funftioniert nur wenn options(knitr.table.format = "latex")
gesetzt wird (das gilt nur fuer Pdf)
}
\details{
markdown: knitr::kable -> kableExtra::kable_styling

Text: knitr::kable

Word: knitr::kable(x, format = "pandoc")

Output_word() flextable::regulartable
}
\examples{
library(gtsummary)
 tbl_summary_ex1 <-
trial \%>\%
  select(age, grade, response) \%>\%
  tbl_summary() \%>\%
  Output()


df1 <- data.frame(
term = c("A", "B", "C", "D"),
n = c(23, 14, 56, 2),
m = c("4.7 (2.4)", "4.1 (2.3)", "8.9 (3.6)", NA)
)


df2 <- data.frame(
  term = c("A", "B", "C", "D"),
  G1_k_n = c(23, 14, 56, 2),
  G1_k_m = c("4.7 (2.4)", "4.1 (2.3)", "8.9 (3.6)", NA),
  G2_n = c(33, 35, 78, 21),
  G2_m = c("4.9 (2.7)", "4.7 (2.5)", "4.1 (5.6)", "4.2 (5.6)")
  )
#+ df-default ,  results='asis'
df1 \%>\% Output()

#+ df-false ,  results='asis'
df1 \%>\% Output()


#+ df-html
df1 \%>\% Output(output="html")

#+ df-text ,  results='asis'
df1 \%>\% Output(output="text")

#+ df-word ,  results='asis'
 # df1 \%>\% Output(output="word")


#+ df-mark
df1 \%>\% Output(output="markdown")

# df1 \%>\%  Output(linesep = c("", "", "\\midrule"))



m <- data.frame(
  Item = factor(1:3, 1:3, c("A", "B", "C")),
  a = (1:3),
  b = (1:3) * 2,
  c = (1:3) * 3
)

m <- stp25tools::Label(m, a = "Aplha", b = "Barbar", c = "Ciklyn")
Text(m)
}
