% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_word.R
\name{Output_word}
\alias{Output_word}
\title{Output nach docx}
\usage{
Output_word(
  x,
  caption = NULL,
  note = NULL,
  output = "docx",
  split_header = TRUE,
  ...
)
}
\arguments{
\item{x}{data.frame}

\item{caption, note, output, split_header}{Ueberschrift usw}

\item{...}{not used}
}
\value{
flextable
}
\description{
Geht nur direkt asis_output() function only works in top-level R expressions,
}
