% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projekt.R
\name{Projekt}
\alias{Projekt}
\alias{Rmd_Start}
\alias{End}
\alias{End2}
\alias{Stop}
\title{Projekt}
\usage{
Projekt(myformat = "", Projektname = "Demo", datum = date())

Rmd_Start(myformat, Projektname, datum)

End(
  browser = NA,
  output = options()$prompt[1] == "HTML> ",
  file = HTMLGetFile()
)

End2(
  browser = NA,
  output = options()$prompt[1] == "HTML> ",
  file = HTMLGetFile(),
  output.dir = file.path(getwd(), get_opt("html_folder"))
)

Stop()
}
\arguments{
\item{myformat}{HTML, Spin, Knit, Rpres oder Text. Spin ist knitr wobei die Ausgabe der Tabellen mit html erfolgt}

\item{Projektname}{Bezeichnung des Projektes (gilt auch fuer die HTML Seite)}

\item{datum}{Datum zur Dokumentation}

\item{browser}{Ie oder Chrome}

\item{output, file}{intern}
}
\value{
html-Pfad oder Projektname als Text
}
\description{
Ein R-File ruft die Interne Funktion Projekt_html() auf und ein Rmd-File Rmd_Start().

Einstellungen fuer .Rmd files hier
werden keine Folder erstellt

\subsection{End}{
Zuruecksetzen der Einstellungen und Aufruf des Browsers browser = getOption("browser")}

Die Funktion geht nicht unter Linux/Apple
}
\details{
Start eines neuen Prpjekts
}
\examples{

\dontrun{


#' Ändern des Speicher-Orts für HTML
 set_opt(html_folder ="Temp-Ordner")

 Projekt()

 # Auswertung

 End()

}
}
