% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SavePlot.R
\name{SavePlot}
\alias{SavePlot}
\alias{SaveGTplot}
\alias{HTML_IMG}
\title{SavePlot}
\usage{
SavePlot(
  caption = "",
  w = grDevices::dev.size("in")[1],
  h = grDevices::dev.size("in")[2],
  filename = "",
  save_plot = "pdf",
  output = which_output(),
  res = 72,
  out.type = "pdf"
)

SaveGTplot(
  x,
  caption = "",
  filename = "",
  save_plot = "pdf",
  output = which_output(),
  res = 72,
  ...
)

HTML_IMG(file, caption = file, size = "50\%", output = which_output())
}
\arguments{
\item{caption}{TextAbbildung}

\item{w, h, res}{Width, Height, an HTMLplot w and h in Inches or Pixel
default = 520, Breite der  Grafik  also zb w=8=dev.size("in")[2],
h=dev.size("in")[1] Hoehe der Grafik also zb h=8
res=72}

\item{filename}{Text}

\item{save_plot}{c("pdf", "esp", "png", "jpeg", "bmp") speichern als file}

\item{output}{nur wichtig bei html}

\item{out.type}{name of the output device: can
be "pdf", or "cairo" dev.copy2pdf}
}
\value{
file name
}
\description{
Speichern von Grafiken
Grafik werden  als PDF oder Windows-Metafile
speichern. Die Funktion Abb() beinhaltet den Pfad.
}
\examples{

\dontrun{

plot(1)
# SavePlot("Einfache Grafik", w=4, h=4)

#'  require(cowplot)
title <-
  cowplot::ggdraw() +
  cowplot::draw_label("FFQ DEGS 534", fontface = 'bold')

cowplot::plot_grid(
  title, NULL,
  function(x) plot(1),
  function(x) plot(1:10),
  nrow = 2,
  #labels = c("", "", "C"),
  rel_heights = c(0.05, 1 )
)
#  ?cowplot::save_plot( )
#ggplot2::ggsave("test-23.pdf")
#ggplot2::ggsave("test-2.eps")

}

#save_my_plot <- function(p,
#                         w = 7,
#                         h = 3.5 ,
#                         filename = "A-Figure.jpeg",
#                         out.type = "cairo",
#                         path = getwd(),
#                         res= 72
#) {
#
 # p_new <-
 #   update(p, par.settings = stp25settings::bw_theme(farbe()))
 # filename <-   file.path(path, filename)
 # jpeg(filename,
 #      width = round(w * res),
 #      height = round(h * res),
 #      quality = 100,
 #      res = res,
#       type  = out.type
 # )
 # # wegen "cairo"
#  print(cowplot::plot_grid(p_new))
 # dev.off()
#
 # (p_new)
#}


# require(lattice)
# require(stp25output2)
# require(stp25settings)
# p<- barchart(yield ~ variety | site, data = barley,
#             groups = year, layout = c(1,6), stack = TRUE,
#             auto.key = list(space = "right"),
#             ylab = "Barley Yield (bushels/acre)",
#             scales = list(x = list(rot = 45)))
#
#   save_my_plot(p)


}
