% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SavePlot.R, R/tab_abb.R
\name{Abb_Index}
\alias{Abb_Index}
\alias{Abb}
\alias{Tab_Abb}
\alias{Tab}
\alias{Tab_Index}
\alias{Caption}
\alias{Note}
\title{Tabellen und Abbildungs Nummern}
\usage{
Abb_Index(x = NULL)

Abb(filename = "", caption = "", folder = get_opt("fig_folder"))

Tab(caption = "", caption.nr.prefix = get_opt("table", "caption.nr.prefix"))

Tab_Index(x = NULL)

Caption(
  caption = NULL,
  atr = NULL,
  N = NULL,
  include.n = get_opt("table", "include.n"),
  include.tabel.number = get_opt("table", "include.tabel.number")
)

Note(x = NULL, atr = NULL)
}
\arguments{
\item{x}{Text}

\item{filename}{Name des Foles}

\item{caption}{}

\item{folder}{Speicherort}

\item{atr}{in Caption: alternativer Text}

\item{include.n, include.tabel.number}{wird über get_opt gesteuert
entweder direkt include.n oder indirekt mit caotion =TRUE/FALSE}
}
\value{
Name plus Nummer
}
\description{
Nummerierte Abbildung default Folder ist fig

Ueberschrift aus stp-Objekt:
\dontrun{
Caption(caption, attr(x, "caption"))
oder Note(note, attr(x, "note")) mit
}
}
