% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SavePlot.R, R/tab_abb.R
\name{Abb_Index}
\alias{Abb_Index}
\alias{Abb}
\alias{Tab_Abb}
\alias{Tab}
\alias{Tab_Index}
\alias{Caption}
\alias{Note}
\title{Tabellen und Abbildungs Nummern}
\usage{
Abb_Index(x = NULL)

Abb(filename = "", caption = "", folder = get_opt("fig_folder"))

Tab(caption = "")

Tab_Index(x = NULL)

Caption(caption = NULL, atr = NULL, N = NULL, include.n = get_opt("caption"))

Note(x = NULL, atr = NULL)
}
\arguments{
\item{x}{Text}

\item{filename}{Name des Foles}

\item{caption}{}

\item{folder}{Speicherort}

\item{atr}{in Caption: alternativer Text}
}
\value{
Name plus Nummer
}
\description{
Nummerierte Abbildung default Folder ist fig

Ueberschrift aus stp-Objekt:

Caption(caption, attr(x, "caption"))
oder Note(note, attr(x, "note")) mit
}
