% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Text.R
\name{Text}
\alias{Text}
\alias{Text.default}
\alias{Text.character}
\alias{Text.data.frame}
\alias{Head}
\alias{Head1}
\alias{Head2}
\alias{Head3}
\alias{Head4}
\alias{Head5}
\alias{Anmerkung}
\alias{Kunde}
\alias{citation_library}
\alias{Zitat}
\alias{Arbeitszeit}
\title{Text Ausgabe}
\usage{
Text(x, ...)

\method{Text}{default}(x, ...)

\method{Text}{character}(..., style = 0, char = "-", output = which_output())

\method{Text}{data.frame}(
  x,
  ...,
  style = 0,
  char = "-",
  include.levels = FALSE,
  output = which_output()
)

Head(..., style = 3, char = "-")

Head1(...)

Head2(...)

Head3(...)

Head4(...)

Head5(...)

Anmerkung(
  ...,
  prafix = "Anmerkung",
  include.comment = if (is.null(get_opt("include.comment"))) TRUE else
    get_opt("include.comment")
)

Kunde(
  x = "",
  msg = NULL,
  name = NULL,
  include.comment = if (is.null(get_opt("include.comment"))) TRUE else
    get_opt("include.comment")
)

citation_library(..., style = "text", .bibstyle = NULL, output = "text")

Zitat(...)

Arbeitszeit(Lines, sep = "\\\\t", ...)
}
\arguments{
\item{x}{Objekt dataframe, liste, vector}

\item{...}{namen der Librarys}

\item{style}{0 = p oder normaler Text 2 = h2 also Ueberschrift bei
Consolen-Ausgabe}

\item{char}{ist nicht zu Verwenden Text Trennzeichen bei Consolen-Ausgabe}

\item{output}{html oder text}

\item{include.levels}{levels mit ausgeben default = FALSE}

\item{prafix}{Anmerkung}

\item{include.comment}{logical.}

\item{msg, name}{Name des kunden oder Nachricht}

\item{.bibstyle}{an format}

\item{Lines}{in Arbeitszeit der Input-String}

\item{sep}{in Arbeitszeit das Input-String-Trennzeichen}
}
\value{
Vector

character
}
\description{
Ausgabe von \code{Text()} und Ueberschriften. \code{Head} ist dabei eine
Kopie von \code{Text} mit
dem Parameter \code{style = 2}

Anmerkung() ist ein blauer Text.

Kunde() ist ein rot-brauner Text.

Zitat() ist eine Text  vom Type <blockquote>

Arbeitszeit() Tabelle zur Dokumentation der Arbeiszeit.
}
\examples{

\dontrun{
#library(stp25)
#Projekt("html")
Head("Ueberschrift") ## HR=2

Text("hallo", "ich bin ein Test")
###plot(sin, -pi, 2*pi,main="Sinus")
###HTMLplot( Caption="Look at this curve!")
#End()
}
\dontrun{

m <- data.frame(
  Item = factor(1:3, 1:3, c("A", "B", "C")),
  a = (1:3),
  b = (1:3) * 2,
  c = (1:3) * 3
)

m <- stp25tools::Label(m, a = "Aplha", b = "Barbar", c = "Ciklyn")
Text(m)
}
\dontrun{
citation_library(base,Hmisc,car,lattice)
}
}
