% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Text.R
\name{Text}
\alias{Text}
\alias{Head}
\alias{Head1}
\alias{Head2}
\alias{Head3}
\alias{Head4}
\alias{Head5}
\alias{Anmerkung}
\alias{Kunde}
\alias{citation_library}
\alias{Zitat}
\alias{Arbeitszeit}
\title{Text Ausgabe}
\usage{
Text(x, ...)

Head(..., style = 3, char = "-")

Head1(...)

Head2(...)

Head3(...)

Head4(...)

Head5(...)

Anmerkung(..., prafix = "Anmerkung")

Kunde(x = "", msg = NULL, name = NULL)

citation_library(..., style = "text", .bibstyle = NULL, output = "text")

Zitat(...)

Arbeitszeit(Lines, sep = "\\\\t", ...)
}
\arguments{
\item{...}{namen der Librarys}

\item{style}{0 = p oder normaler Text 2 = h2 also Ueberschrift bei
Consolen-Ausgabe}

\item{char}{ist nicht zu Verwenden Text Trennzeichen bei Consolen-Ausgabe}

\item{prafix}{Anmerkung}

\item{output}{html oder text}

\item{Lines}{in Arbeitszeit der Input-String}

\item{sep}{in Arbeitszeit das Input-String-Trennzeichen}

\item{style, .bibstyle}{an format}
}
\value{
Vector

character
}
\description{
Ausgabe von \code{Text()} und Ueberschriften. \code{Head} ist dabei eine
Kopie von \code{Text} mit
dem Parameter \code{style = 2}

Anmerkung() ist ein blauer Text.

Kunde() ist ein rot-brauner Text.

Zitat() ist eine Text  vom Type <blockquote>

Arbeitszeit() Tabelle zur Dokumentation der Arbeiszeit.
}
\examples{

\dontrun{
#library(stp25)
#Projekt("html")
Head("Ueberschrift") ## HR=2

Text("hallo", "ich bin ein Test")
###plot(sin, -pi, 2*pi,main="Sinus")
###HTMLplot( Caption="Look at this curve!")
#End()
}
\dontrun{
citation_library(base,Hmisc,car,lattice)
}
}
